Moves the email connector send method to the email connector file.
