+++ src/fidesops/ops/service/connectors/email_connector.py
@@ -1,6 +1,7 @@
 import logging
 from typing import Any, Dict, List, Optional
 
+from fideslib.models.audit_log import AuditLog, AuditLogAction
 from sqlalchemy.orm import Session
 
 from fidesops.ops.common_exceptions import (
@@ -9,7 +10,12 @@ from fidesops.ops.common_exceptions import (
 )
 from fidesops.ops.graph.config import CollectionAddress, FieldPath
 from fidesops.ops.graph.traversal import TraversalNode
-from fidesops.ops.models.connectionconfig import ConnectionTestStatus
+from fidesops.ops.models.connectionconfig import (
+    ConnectionConfig,
+    ConnectionTestStatus,
+    ConnectionType,
+)
+from fidesops.ops.models.datasetconfig import DatasetConfig
 from fidesops.ops.models.policy import CurrentStep, Policy, Rule
 from fidesops.ops.models.privacy_request import (
     CheckpointActionRequired,
@@ -84,7 +90,7 @@ class EmailConnector(BaseConnector[None]):
     ) -> Optional[List[Row]]:
         """Access requests are not supported at this time."""
         logger.info(
-            "Access request not supported for email connector `%s` at this time.",
+            "Access requests not supported for email connector '%s' at this time.",
             node.address.value,
         )
         return []
@@ -112,6 +118,8 @@ class EmailConnector(BaseConnector[None]):
             action_needed=[manual_action],
         )
 
+        # Raises a special exception just to update the ExecutionLog message.  The email send itself
+        # is postponed until all collections have been visited.
         raise PrivacyRequestErasureEmailSendRequired("email prepared")
 
     def build_masking_instructions(
@@ -148,3 +156,62 @@ class EmailConnector(BaseConnector[None]):
         # Returns a ManualAction even if there are no fields to mask on this collection,
         # because the locators still may be needed to find data to mask on dependent collections
         return ManualAction(locators=locators, update=mask_map if mask_map else None)
+
+
+def email_connector_erasure_send(db: Session, privacy_request: PrivacyRequest) -> None:
+    """
+    Send emails to configured third-parties with instructions on how to erase remaining data.
+    Combined all the collections on each email-based dataset into one email.
+    """
+    email_dataset_configs = db.query(DatasetConfig, ConnectionConfig).filter(
+        DatasetConfig.connection_config_id == ConnectionConfig.id,
+        ConnectionConfig.connection_type == ConnectionType.email,
+    )
+    for ds, cc in email_dataset_configs:
+        template_values: Dict[
+            str, Optional[CheckpointActionRequired]
+        ] = privacy_request.get_email_connector_template_contents_by_dataset(
+            CurrentStep.erasure, ds.dataset.get("fides_key")
+        )
+
+        if not template_values:
+            logger.info(
+                "No email sent: no template values saved for '%s'",
+                ds.dataset.get("fides_key"),
+            )
+            return
+
+        if not any(
+            (
+                action_required.action_needed[0].update
+                if action_required and action_required.action_needed
+                else False
+                for action_required in template_values.values()
+            )
+        ):
+            logger.info(
+                "No email sent: no masking needed on '%s'", ds.dataset.get("fides_key")
+            )
+            return
+
+        dispatch_email(
+            db,
+            action_type=EmailActionType.EMAIL_ERASURE_REQUEST_FULFILLMENT,
+            to_email=cc.secrets.get("to_email"),
+            email_body_params=template_values,
+        )
+
+        logger.info(
+            "Email send succeeded for request '%s' for dataset: '%s'",
+            privacy_request.id,
+            ds.dataset.get("fides_key"),
+        )
+        AuditLog.create(
+            db=db,
+            data={
+                "user_id": "system",
+                "privacy_request_id": privacy_request.id,
+                "action": AuditLogAction.email_sent,
+                "message": f"Erasure email instructions dispatched for '{ds.dataset.get('fides_key')}'",
+            },
+        )
