+++ src/fidesops/ops/service/privacy_request/request_runner_service.py
@@ -23,7 +23,7 @@ from fidesops.ops.graph.analytics_events import (
     fideslog_graph_failure,
 )
 from fidesops.ops.graph.graph import DatasetGraph
-from fidesops.ops.models.connectionconfig import ConnectionConfig, ConnectionType
+from fidesops.ops.models.connectionconfig import ConnectionConfig
 from fidesops.ops.models.datasetconfig import DatasetConfig
 from fidesops.ops.models.policy import (
     ActionType,
@@ -34,13 +34,11 @@ from fidesops.ops.models.policy import (
     WebhookTypes,
 )
 from fidesops.ops.models.privacy_request import (
-    CheckpointActionRequired,
     PrivacyRequest,
     PrivacyRequestStatus,
     can_run_checkpoint,
 )
-from fidesops.ops.schemas.email.email import EmailActionType
-from fidesops.ops.service.email.email_dispatch_service import dispatch_email
+from fidesops.ops.service.connectors.email_connector import email_connector_erasure_send
 from fidesops.ops.service.storage.storage_uploader_service import upload
 from fidesops.ops.task.filter_results import filter_data_categories
 from fidesops.ops.task.graph_task import (
@@ -399,62 +397,3 @@ def generate_id_verification_code() -> str:
     Generate one-time identity verification code
     """
     return str(random.choice(range(100000, 999999)))
-
-
-def email_connector_erasure_send(db: Session, privacy_request: PrivacyRequest) -> None:
-    """
-    Send emails to configured third-parties with instructions on how to erase remaining data.
-    Combined all the collections on each email-based dataset into one email.
-    """
-    email_dataset_configs = db.query(DatasetConfig, ConnectionConfig).filter(
-        DatasetConfig.connection_config_id == ConnectionConfig.id,
-        ConnectionConfig.connection_type == ConnectionType.email,
-    )
-    for ds, cc in email_dataset_configs:
-        template_values: Dict[
-            str, Optional[CheckpointActionRequired]
-        ] = privacy_request.get_email_connector_template_contents_by_dataset(
-            CurrentStep.erasure, ds.dataset.get("fides_key")
-        )
-
-        if not template_values:
-            logger.info(
-                "No email sent: no template values saved for '%s'",
-                ds.dataset.get("fides_key"),
-            )
-            return
-
-        if not any(
-            (
-                action_required.action_needed[0].update
-                if action_required and action_required.action_needed
-                else False
-                for action_required in template_values.values()
-            )
-        ):
-            logger.info(
-                "No email sent: no masking needed on '%s'", ds.dataset.get("fides_key")
-            )
-            return
-
-        dispatch_email(
-            db,
-            action_type=EmailActionType.EMAIL_ERASURE_REQUEST_FULFILLMENT,
-            to_email=cc.secrets.get("to_email"),
-            email_body_params=template_values,
-        )
-
-        logger.info(
-            "Email send succeeded for request '%s' for dataset: '%s'",
-            privacy_request.id,
-            ds.dataset.get("fides_key"),
-        )
-        AuditLog.create(
-            db=db,
-            data={
-                "user_id": "system",
-                "privacy_request_id": privacy_request.id,
-                "action": AuditLogAction.email_sent,
-                "message": f"Erasure email instructions dispatched for '{ds.dataset.get('fides_key')}'",
-            },
-        )
